refactor hook to util
