+++ protocol-designer/src/pages/Designer/ProtocolSteps/StepForm/PipetteFields/TipSelectionWizard/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './useMemoizedTipAccessibilityByTiprackIdByWellName'
-export * from './useValidTiprackIds'
