+++ protocol-designer/src/pages/Designer/ProtocolSteps/StepForm/PipetteFields/TipSelectionWizard/utils.ts
@@ -7,6 +7,8 @@ import {
 } from '@opentrons/shared-data'
 import {
   COLUMN_4_SLOTS,
+  getIsSafePickupWithinTiprack,
+  getIsSafePipetteMovement,
   getSlotInLocationStack,
 } from '@opentrons/step-generation'
 
@@ -17,7 +19,11 @@ import type {
   PipetteChannels,
   PipetteV2Specs,
 } from '@opentrons/shared-data'
-import type { LabwareEntities } from '@opentrons/step-generation'
+import type {
+  InvariantContext,
+  LabwareEntities,
+  TimelineFrame,
+} from '@opentrons/step-generation'
 import type {
   AllTemporalPropertiesForTimelineFrame,
   LabwareOnDeck,
@@ -179,4 +185,98 @@ export const getAffectedWells = (args: {
     return Object.keys(labwareDef.wells)
   }
   return []
+}
+
+export const getValidTiprackIds = (args: {
+  pipetteId: string
+  nozzles: NozzleConfigurationStyle
+  channels: PipetteChannels
+  numPickups: number
+  primaryNozzle: string
+  invariantContext: InvariantContext
+  robotState: TimelineFrame | null
+  tipAccessibilityStatus: Record<string, Record<string, boolean>>
+}): string[] => {
+  const {
+    pipetteId,
+    nozzles,
+    channels,
+    numPickups,
+    primaryNozzle,
+    tipAccessibilityStatus,
+    invariantContext,
+    robotState,
+  } = args
+  const { labwareEntities } = invariantContext
+  const validTiprackIds = Object.keys(tipAccessibilityStatus).reduce<string[]>(
+    (acc, id) => {
+      const tiprackDef = labwareEntities[id].def
+      const tipState = robotState?.tipState.tipracks[id] ?? {}
+      const stack = robotState?.labware[id]?.stack
+      const isPickupCompatibleWithPossibleAdapter =
+        stack != null
+          ? getIsPickupCompatibleWithPossibleAdapter(
+              stack,
+              labwareEntities,
+              nozzles,
+              channels
+            )
+          : true
+      if (!isPickupCompatibleWithPossibleAdapter) {
+        return acc
+      }
+
+      let isValidTiprack: boolean = true
+      const addedWells: string[] = []
+      for (let pickupIndex = 0; pickupIndex < numPickups; pickupIndex++) {
+        const wellsToTraverse = Object.keys(tiprackDef.wells)
+        let foundSafePickup = false
+        for (const wellName of wellsToTraverse) {
+          const { isSafe: isSafeWithinTiprack, isComplete } =
+            getIsSafePickupWithinTiprack({
+              tipState,
+              primaryNozzle,
+              channels,
+              nozzleConfiguration: nozzles,
+              wellName,
+              tiprackDef: tiprackDef,
+              tipsToIgnore: addedWells,
+            })
+          const isSafeMoveConsideringDeck =
+            robotState != null
+              ? getIsSafePipetteMovement({
+                  robotState,
+                  invariantContext,
+                  pipetteId,
+                  labwareId: id,
+                  wellTargetName: wellName,
+                  primaryNozzle,
+                  nozzleConfiguration: nozzles,
+                })
+              : true
+          if (isSafeWithinTiprack && isSafeMoveConsideringDeck && isComplete) {
+            const allAffectedWells = getAffectedWells({
+              wellName,
+              labwareDef: tiprackDef,
+              channels,
+              nozzles,
+            })
+            addedWells.push(...allAffectedWells)
+            foundSafePickup = true
+            break // Found a safe pickup for this iteration, move to next pickup
+          }
+        }
+
+        // If we didn't find a safe pickup for this iteration, the tiprack is invalid
+        if (!foundSafePickup) {
+          isValidTiprack = false
+          break // Stop checking this tiprack entirely
+        }
+      }
+
+      return isValidTiprack ? [...acc, id] : acc
+    },
+    []
+  )
+  return validTiprackIds
 }
